+++ src/main/scala/com/ironcorelabs/DeviceName.scala
@@ -1,0 +1,18 @@
+package com.ironcorelabs.scala.sdk
+
+import cats.effect.Sync
+import com.ironcorelabs.{sdk => jsdk}
+
+/**
+ * Device name type. Validates that the provided name isn't an empty string
+ */
+case class DeviceName(name: String) {
+  private[sdk] def toJava[F[_]](implicit syncF: Sync[F]): F[jsdk.DeviceName] =
+    syncF.delay(jsdk.DeviceName.validate(name))
+}
+
+object DeviceName {
+  def apply(name: jsdk.DeviceName): Option[DeviceName] =
+    Option(name).map(n => DeviceName(n.getName))
+  // def apply(name: String): DeviceName = DeviceName(Some(name))
+}