+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update to ironoxide-java 0.7.0
 - Add `groupAddMembers()` and `groupRemoveMembers()` to add/remove group members.
 - Add `groupAddAdmins()` and `groupRemoveAdmins()` to add/remove group admins.
+- Add `generateNewDevice()` to generate a new device for the user.
 
 ## 0.5.2